requirements.txt
fastapi
python-dotenv
uvicorn
langgraph
langchain
langchain[google-genai]
pyTelegramBotAPI
requests
pydantic